Update RestaurantService.java
updated searchRestaurant and it's helping methods. Tries to search off of one query input and then determine if the query is a float, integer, or string to then determine if it is search by rating, price, or name/cuisine. But right now it cannot differentiate between a float and int. Added restaurant update method to update name, address. phonenumber, hours, and description, but not photos yet